anomaly-detection/requirements.txt
pika==1.3.2
python-dotenv==1.0.0
loguru==0.7.3